Fetch Metadata: Add a test for `<embed>` navigated after loading.
@annevk noted in https://github.com/whatwg/fetch/pull/948#discussion_r331921654
that it was possible to navigate an `<embed>` after it loads. In this
case, it seems safest to ensure that this navigation shows up with a
destination of `embed` and a mode of `no-cors`.

That seems to be what we're sending today via Fetch Metadata, so let's
lock it in via this patch's tests. I suspect it's not what the
underlying `fetch()` API would show in a service worker, but we can
deal with that in a separate patch.

Bug: 1011763
Change-Id: I0156b2b1b467c914b15f6af50dfa37ce74db6c62