Portals: Add WPTs for navigations in portals
We test that methods for history manipulation work inside a portal
context and that their behaviour is the same as in a non-portal context.
We also test the new behaviour introduced with portals where a portal
and its host do not form a joint session history. So unlike iframes,
any session history change within a portal is done independently of its
host.

Bug: 914108
Change-Id: Ib1cdbe3e8d1c0e009eea1f05af28ef2971e8b353