NG/ContentSize: Save unconstrained intrinsic block size in NG for overflow calc.
This patch saves the unconstrained intrinsic box size during the
block layout algorithm in order to use it for overflow calculation
in the node. Since we can override the intrinsic size with css-content-size,
we may cause scrollbars to appear. By using unconstrained size, we retain
correct behavior in which scrollbars only depend on the child content,
and not on content-size override.

Bug: 991096
R=chrishtr@chromium.org, ikilpatrick@chromium.org
Change-Id: I0d9cb270ee710b759f93202497f69bf72544f81a