[css-text] Consider breaking opportunities of inline siblings
First of all, bear in mind that this change affects only to
intrinsic size computation in legacy layout.

During the intrinsic size computation of blocks with inline children,
we are determining the breaking opportunities for each child. When the
word-break: break-word is used, each of this breaking opportunities
should account for the min-content size.

Hence, we can't simply ignore that previous inline siblings had an
opportunity already. This patch proposes to consider that both values,
break-word and break-all, imply that every position is a potential
breakable location, at least regarding the intrinsic size computation.

Bug: 1013644
Change-Id: I04261b323cd029a624724363a566fccc826863af