wake lock: Implement new proposed API.
This is another big update to the implementation, which now follows the
API proposed in https://github.com/w3c/wake-lock/issues/226.

From a Blink/Chromium internals perspective, things have not changed much,
and we still perform the same permission checks and connect to the wake lock
service via WakeLockStateRecord. The biggest change is that WakeLockController
was merged into WakeLock itself.

From an end-user perspective, though, the API now looks like a mix of the
old, navigator-based API and the one we implemented a few months ago:

* The main idea is that WakeLock.request() no longer returns a promise that
  does not resolve while the lock is held, as that is not very intuitive.
  Instead, it returns a new WakeLockSentinel object.
* WakeLockSentinel is a new interface that can release() a lock and receive
  "release" events.
* We do go back to a Navigator-based API, but we also support
  WorkerNavigator.
* There is no WakeLockRequest interface or state tracking in the WakeLock
  interface.
* WakeLock.requestPermission() was removed.
* Consequently, we now have a WakeLockEvent interface.

The changes to the spec still need to land, but there is some rough
consensus on what the API should look like, and we do not want to miss the
M79 branch cut.

Bug: 257511, 1010162
Change-Id: If1baed914ccd9eb7103a1bb71a40529976777ec2